Smooth out bubbles fading
When the bubbles stop, we let them run their animation 3 times before
removing them. If we just change the iteration-count from "infinite" to
3, then the animation will stop abruptly if it already has run more than
3 iterations while "infinite". There is no easy way to know how many
iterations an animation has run, so we need to keep track of it
manually, via a data attribute (not sure what the perfs implications of
that are).